requirements.txt
faiss-cpu
sentence-transformers
spacy
pandas
python-dotenv
numpy
